Organize add student feature into a parallel route
