--- src/app/course/[id]/@students/AddStudentModal/index.ts
@@ -1,4 +1,0 @@
-import AddStudentModal from './AddStudentModal';
-
-export default AddStudentModal;
-
